inventory_view: splitting controller into sub-libs
